Posterior Predictive Probabilities (PPPs)

Computes posterior predictive probabilities (PPPs) based on the odds ratios for each pair of items.

PPP(object, ...)

# S3 method for edina
PPP(object, alpha = 0.05, ...)

Arguments

object
An edina object

...
Not used.

alpha
Defining region to indicate the level of extremeness the data must before the model is problematic.

Value

The PPP value given the specified alpha value.

Details

PPPs that smaller than 0.05 or greater than 0.95 tend to be extreme and evidence of misfit. As a result, this is more of a heuristic metric.

PPP Computation Procedure

simulate observed responses \(\mathbf Y^{(r)}\) using model parameters from iteration \(r\) of the MCMC sampler

computing the odds ratio for each pair of items at iteration \(r\) as $$OR^{(r)} = n_{11}^{(r)}n_{00}^{(r)}/\left(n_{10}^{(r)}n_{01}^{(r)}\right)$$, where \(n_{11}^{(r)}\) is the frequency of ones on both variables at iteration \(r\), \(n_{10}^{(r)}\) is the frequency of ones on the first item and zeros on the second at iteration \(r\), etc.; and

computing PPPs for each item pair as the proportion of generated \(OR^{(r)}\)'s that exceeded elements of the observed odds ratios.